+++ HISTORY.md
@@ -1,6 +1,7 @@
 # Rocksdb Change Log
 ## Unreleased
 ### Public API changes
+* Add TransactionLogIterator::RecycleWriteBatch() method to optionally recycle a WriteBatch handed out by TransactionLogIterator::GetBatch(). This is an optional API that can be used to avoid repeated memory allocations for new WriteBatch objects when calling GetBatch() on the same TransactionLogIterator instance. Using the API is performance optimization which should not affect correctness.
 * Add new API GetUnixTime in Snapshot class which returns the unix time at which Snapshot is taken.
 * Add transaction `get_pinned` and `multi_get` to C API.
 * Add two-phase commit support to C API.
@@ -16,6 +17,9 @@
 
 ### Behavior changes
 * DB::Open(), DB::OpenAsSecondary() will fail if a Logger cannot be created (#9984)
+
+### Performance Improvements
+* Embedders can use the new TransactionLogIterator::RecycleWriteBatch() method to optionally return a WriteBatch handed out by TransactionLogIterator::GetBatch() back to the TransactionLogIterator. This is an optional API that can be used to avoid repeated memory allocations for new WriteBatch objects when calling GetBatch() on the same TransactionLogIterator instance.
 
 ## 7.3.0 (05/20/2022)
 ### Bug Fixes